+++ db/migrate/20100607020132_create_venda.rb
@@ -4,14 +4,14 @@ class CreateVenda < ActiveRecord::Migration
       t.references :cliente
       t.references :vendedor
 
-      t.date   :data
-      t.string :tipo,  :limit => 1
-      t.float  :valor, :scale=>2, :precision=>9
-      t.string :obs,   :limit => 100
+      t.date     :data
+      t.string   :tipo,  :limit => 1
+      t.decimal  :valor, :scale=>2, :precision=>9
+      t.string   :obs,   :limit => 100
 
       t.timestamps
     end
-    
+
     add_index :vendas, :cliente_id
     add_index :vendas, :vendedor_id
