+++ db/migrate/20100607020143_create_venda_item.rb
@@ -4,13 +4,13 @@ class CreateVendaItem < ActiveRecord::Migration
       t.references :venda,    :null => false
       t.references :item,     :null => false
 
-      t.float      :bruto,    :scale=>2, :precision=>9
-      t.float      :desconto, :scale=>2, :precision=>4
-      t.float      :valor,    :scale=>2, :precision=>9
+      t.decimal      :bruto,    :scale=>2, :precision=>9
+      t.decimal      :desconto, :scale=>2, :precision=>4
+      t.decimal      :valor,    :scale=>2, :precision=>9
 
       t.timestamps
     end
-    
+
     add_index :venda_itens, :venda_id
     add_index :venda_itens, :item_id
   end
